requirements.txt
# --------- pytorch --------- #
torch
lightning
torchvision
torchaudio
torchtext
transformers
datasets
opencv-python
torchmetrics

# --------- Basic --------- #
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn

# --------- hydra --------- #
hydra-core
hydra-colorlog
hydra-optuna-sweeper

# --------- Advanced --------- #
salesforce-lavis
open_clip_torch
peft

# --------- loggers --------- #
wandb

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh            # for running bash commands in some tests (linux/macos only)
h5py
filelock

# --------- jupyter --------- #
ipykernel       # for jupyter notebook
jupyter         # for jupyter notebook
pyqt5
pyqtwebengine
